Script now stops X11 only when all the containers are stopped

+++ windows/build.py
@@ -121,12 +121,13 @@ subprocess.run(command, shell=True)
 ########################################
 
 # Save ipconfig results into file
-command = "ipconfig /all > ipconfig.txt"
+filename = ".ipconfig"
+command = f"ipconfig /all > {filename}"
 subprocess.run(command, shell=True)
 # Get the X11 server IPv4 address from file
-ip = get_xserv_ip('ipconfig.txt')
+ip = get_xserv_ip(filename)
 # Remove file
-os.remove("ipconfig.txt")
+os.remove(filename)
 # Check if IP was found
 if ip is None:
     # IP was not found, exit
@@ -161,10 +162,14 @@ command = f'docker run --rm --name {sys.argv[1]} -it -P --volume="{oz_dir_host}:
 subprocess.run(command, shell=True)
 
 
-#############################
-# CLEANING: Stop X11 server #
-#############################
+######################################################################################
+# CLEANING: Stop X11 server, if all the instances of the container have been stopped #
+######################################################################################
 
-print("Stopping X11 server.")
-command = "taskkill /f /im vcxsrv.exe"
-subprocess.run(command, shell=True)
+command = f"docker ps -aq -f ancestor={image}"
+output = subprocess.run(command, shell=True, stdout=subprocess.PIPE)
+if not output.stdout:
+    # Output of Docker list is empty, no more containers
+    print("Stopping X11 server.")
+    command = "taskkill /f /im vcxsrv.exe"
+    subprocess.run(command, shell=True)
